# Conflicts:
#	connectors/mongodb-connector/src/main/java/io/tapdata/mongodb/MongodbConnector.java
#	connectors/mongodb-connector/src/main/java/io/tapdata/mongodb/MongodbTest.java
#	connectors/mongodb-connector/src/main/java/io/tapdata/mongodb/MongodbUtil.java
#	connectors/mysql-connector/src/main/java/io/tapdata/connector/mysql/MysqlConnector.java
#	connectors/selectdb-connector/src/main/java/io/tapdata/connector/selectdb/SelectDbConnector.java
#	connectors/selectdb-connector/src/main/java/io/tapdata/connector/selectdb/SelectDbStreamLoader.java
#	manager/tm-common/src/main/java/com/tapdata/tm/commons/task/dto/TaskDto.java
#	manager/tm/src/main/java/com/tapdata/tm/ds/service/impl/DataSourceService.java
#	manager/tm/src/main/java/com/tapdata/tm/externalStorage/service/ExternalStorageService.java
#	manager/tm/src/main/java/com/tapdata/tm/metadatainstance/service/MetadataInstancesService.java
#	manager/tm/src/main/java/com/tapdata/tm/task/entity/TaskEntity.java
#	manager/tm/src/main/java/com/tapdata/tm/task/service/TaskService.java
#	manager/tm/src/main/java/com/tapdata/tm/task/service/impl/LdpServiceImpl.java
#	tapdata-cli/src/main/java/io/tapdata/pdk/cli/RegisterMain.java

+++ manager/tm/src/main/java/com/tapdata/tm/user/service/UserService.java
@@ -399,8 +399,10 @@ public class UserService extends BaseService<UserDto, User, ObjectId, UserReposi
     }
     private List<RoleMappingDto> updateRoleMapping(String userId, List<Object> roleusers, UserDetail userDetail) {
         // delete old role mapping
-        long deleted = roleMappingService.deleteAll(Query.query(Criteria.where("principalId").is(userId).and("principalType").is("USER")));
-        log.info("delete old role mapping for userId {}, deleted: {}", userId, deleted);
+        if (!userDetail.getEmail().equals("admin@admin.com")) {
+            long deleted = roleMappingService.deleteAll(Query.query(Criteria.where("principalId").is(userId).and("principalType").is("USER")));
+            log.info("delete old role mapping for userId {}, deleted: {}", userId, deleted);
+        }
         // add new role mapping
         if (CollectionUtils.isNotEmpty(roleusers)) {
             List<RoleMappingDto> roleMappingDtos = roleusers.stream().map(r -> (String) r).map(roleId -> {
